if snapshot name already exists and -force specified, safely overwrite it
With previous PR #17 we detected if a snapshot name was already present and refused to proceed, unless `packer build -force` was specified. If it was specified, then we would _add_ the new snapshot to the existing one.

This PR brings the behavior of this plugin completely on par with the expected behavior of `packer build -force`:

    -force    Force a build to continue if artifacts exist, deletes existing artifacts.

The overwite is done safely: first a new image is built and a snapshot is taken. Only if the new snapshot is taken with success we then delete the old one.

As for PR #17, this PR comes with full test coverage.

NOTE: while developing this feature, I noticed two pre-existing bugs, fixed in commits efcc738a4f and d9b9ba31f7. Probably worthwile to involve somebody from Hetzener cloud to double-check.

Closes #14