add pre validate step (check snapshot names) (#17)
Before, it was possible to create multiple snapshots with the same name, leading to a confusing and wasteful situation.

Now, we perform the same kind of checks done by the AWS EC2 builder, and refuse to proceed if there is an existing snapshot with the same name.

Invoking packer build -force will bypass the checks, thus multiple snapshots with the same name will be allowed. This is equivalent to the previous behavior of this plugin.

Part of #14.