feat: add labels options to server and ssh keys (#128)
Continuation of #92 

Adding labels to the temporary server and ssh key objects which are
needed to create the snapshot.

This is needed in some use cases where e.g. packer build stops
unexpectedly due to a SIGKILL and cannot shut down properly. In order to
realize that the respective server and ssh key are orphaned, labels are
needed.

---------

Signed-off-by: janiskemper <janis.kemper@syself.com>
Co-authored-by: janiskemper <janis.kemper@syself.com>